requirements_dev.txt
selenium==2.53.5
mixer==5.5.7
pytest==3.0.0
pytest-django==3.0.0
pytest-html==1.10.0
pytest-cov==2.3.1
